test getallscopes unique values
